single_agent/wandb/run-20260316_225623-wvdfekh8/files/requirements.txt
typing-inspection==0.4.2
Markdown==3.10.2
tensorboard==2.20.0
annotated-types==0.7.0
Jinja2==3.1.6
nvidia-nccl-cu12==2.27.5
typeguard==4.5.1
tensorboard-data-server==0.7.2
pyparsing==3.3.2
sentry-sdk==2.54.0
wandb==0.25.1
pandas==3.0.1
tyro==1.0.9
pydantic==2.12.5
grpcio==1.78.0
cycler==0.12.1
pillow==12.1.1
GitPython==3.1.46
kiwisolver==1.5.0
typing_extensions==4.15.0
certifi==2026.2.25
atari-py==0.2.9
gym==0.23.1
matplotlib==3.10.8
Werkzeug==3.1.6
nvidia-cufile-cu12==1.13.1.3
idna==3.11
cuda-pathfinder==1.4.2
nvidia-cuda-cupti-cu12==12.8.90
smmap==5.0.3
nvidia-nvshmem-cu12==3.4.5
nvidia-curand-cu12==10.3.9.90
stable_baselines3==2.7.1
nvidia-nvtx-cu12==12.8.90
nvidia-cuda-runtime-cu12==12.8.90
cloudpickle==3.1.2
filelock==3.25.2
setuptools==82.0.1
gymnasium==1.2.3
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparselt-cu12==0.7.1
nvidia-nvjitlink-cu12==12.8.93
MarkupSafe==3.0.3
nvidia-cusparse-cu12==12.5.8.93
contourpy==1.3.3
gitdb==4.0.12
protobuf==6.33.5
nvidia-cudnn-cu12==9.10.2.21
python-dateutil==2.9.0.post0
fsspec==2026.2.0
mpmath==1.3.0
platformdirs==4.9.4
pydantic_core==2.41.5
requests==2.32.5
urllib3==2.6.3
networkx==3.6.1
numpy==2.4.3
cuda-bindings==12.9.4
torch==2.10.0
charset-normalizer==3.4.5
nvidia-cufft-cu12==11.3.3.83
opencv-python==4.13.0.92
six==1.17.0
nvidia-cuda-nvrtc-cu12==12.8.93
docstring_parser==0.17.0
gym-notices==0.1.0
psutil==7.2.2
sympy==1.14.0
absl-py==2.4.0
Farama-Notifications==0.0.4
triton==3.6.0
nvidia-cublas-cu12==12.8.4.1
click==8.3.1
fonttools==4.62.1
packaging==26.0
PyYAML==6.0.3
